Adding the bld.bat
Despite being "no arch" a script is still required. Your script is too complex to be integrated to the meta.yaml. That bld.bat (source https://github.com/AnacondaRecipes/conda-feedstock/blob/master/recipe/bld.bat) alllow the build.sh to be executed on Windows. Without that file, the building was incomplete on Windows and that's why there as an error message about having no module called splore.

I tried it twice on my own pc and it did work